# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/user/login": {
            "post": {
                "description": "Post 请求, 发送 multipart/form-data 类型的表单数据, 参数在消息体中",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userApi"
                ],
                "summary": "Post 请求, 发送 multipart/form-data 类型的表单数据, 参数在消息体中",
                "parameters": [
                    {
                        "maxLength": 20,
                        "minLength": 1,
                        "type": "string",
                        "description": "要注册的用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "maxLength": 128,
                        "minLength": 1,
                        "type": "string",
                        "description": "用户密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.HttpOk"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http400"
                        }
                    },
                    "404": {
                        "description": "Page Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http404"
                        }
                    },
                    "500": {
                        "description": "InternalError",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http500"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "Post 请求, 发送 multipart/form-data 类型的表单数据, 参数在消息体中",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userApi"
                ],
                "summary": "Post 请求, 发送 multipart/form-data 类型的表单数据, 参数在消息体中",
                "parameters": [
                    {
                        "maxLength": 20,
                        "minLength": 1,
                        "type": "string",
                        "description": "要注册的用户名",
                        "name": "username",
                        "in": "formData"
                    },
                    {
                        "maxLength": 128,
                        "minLength": 1,
                        "type": "string",
                        "description": "用户密码",
                        "name": "password",
                        "in": "formData"
                    },
                    {
                        "maxLength": 11,
                        "minLength": 11,
                        "type": "string",
                        "description": "电话号码",
                        "name": "tel",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.HttpOk"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http400"
                        }
                    },
                    "404": {
                        "description": "Page Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http404"
                        }
                    },
                    "500": {
                        "description": "InternalError",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http500"
                        }
                    }
                }
            }
        },
        "/user/updatePassword": {
            "post": {
                "description": "Post 请求, 发送 multipart/form-data 类型的表单数据, 参数在消息体中",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userApi"
                ],
                "summary": "Post 请求, 发送 multipart/form-data 类型的表单数据, 参数在消息体中",
                "parameters": [
                    {
                        "maxLength": 20,
                        "minLength": 1,
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "maxLength": 128,
                        "minLength": 1,
                        "type": "string",
                        "description": "用户旧密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "maxLength": 128,
                        "minLength": 1,
                        "type": "string",
                        "description": "用户新密码",
                        "name": "newPassword",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.HttpOk"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http400"
                        }
                    },
                    "404": {
                        "description": "Page Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http404"
                        }
                    },
                    "500": {
                        "description": "InternalError",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http500"
                        }
                    }
                }
            }
        },
        "/user/updateUser": {
            "post": {
                "description": "Post 请求, 发送 multipart/form-data 类型的表单数据, 参数在消息体中",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userApi"
                ],
                "summary": "Post 请求, 发送 multipart/form-data 类型的表单数据, 参数在消息体中",
                "parameters": [
                    {
                        "maxLength": 20,
                        "minLength": 1,
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData"
                    },
                    {
                        "maxLength": 128,
                        "minLength": 1,
                        "type": "string",
                        "description": "用户密码",
                        "name": "password",
                        "in": "formData"
                    },
                    {
                        "maxLength": 11,
                        "minLength": 11,
                        "type": "string",
                        "description": "电话号码",
                        "name": "tel",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.HttpOk"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http400"
                        }
                    },
                    "404": {
                        "description": "Page Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http404"
                        }
                    },
                    "500": {
                        "description": "InternalError",
                        "schema": {
                            "$ref": "#/definitions/swagger.Http500"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "swagger.Http400": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "msg": {
                    "type": "string",
                    "example": "Bad Request"
                }
            }
        },
        "swagger.Http404": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "msg": {
                    "type": "string",
                    "example": "Page Not Found"
                }
            }
        },
        "swagger.Http500": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "msg": {
                    "type": "string",
                    "example": "服务器内部错误"
                }
            }
        },
        "swagger.HttpOk": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "msg": {
                    "type": "string",
                    "example": "接口调用成功"
                }
            }
        }
    }
}